Add an into-SSA pass and an out-of-SSA pass
Working in SSA is way more convenient for dataflow analyses that compute something
for each variable (since the state does not need to be stored per (block,pair) but
per-variable).

Writing the into SSA pass is a mess, but a mess that hopefully we'll only deal
with once. Bite the bullet and do SSA.

At the moment, the compiler does into SSA and then immediately after out of SSA,
because the backend does not understand phi nodes (obviously).