Add support for multiple interfaces of the same type
Interfaces are now given an ID, which is unique per space, and lets
them differentiate who is connected to what to dispatch events.

This required quite a bit of churn in the primitives to keep the
Concrete* classes (formerly Instantiated*) still in the same amount
per space, which avoids duplicated code.

Technically, this also allows multiple instances of the same model
in the same space, but this is explicitly disallowed (as otherwise
the security analysis would probably get very confused).